Revert "Feat account page"
