--- src/components/UserPage/userPage.css
@@ -1,55 +1,0 @@
-.userPage h2 {
-  padding: 1rem 0;
-  font-size: 2rem;
-  font-weight: 500;
-  text-align: center;
-  color: yellow;
-  /*color: #b1091d;*/
-}
-
-.isAllowShowVisited {
-  margin: 2rem auto;
-  text-align: center;
-}
-
-.isAllowShowVisited input, span {
-  display: inline-block;
-  vertical-align: middle;
-}
-
-.isAllowShowVisited input {
-  width: 1.4rem;
-  height: 1.4rem;
-  margin: 0 0.5rem;
-  accent-color: rgba(20, 161, 20, 0.78);
-}
-
-.isAllowShowVisited span {
-  font-size: 1.5rem;
-  color: #eaeaea;
-}
-
-.logoutButton {
-  display: block;
-  margin: 1rem auto;
-  padding: 0.2rem;
-  text-align: center;
-  font-size: 1.5rem;
-  font-weight: 400;
-  color: #61ff10;
-  background: #b1091d;
-  border: 2px solid #b1091d;
-  border-radius: 0.8rem;
-  opacity: 0.8;
-  cursor: pointer;
-  transition: opacity .2s;
-}
-
-.logoutClear {
-  font-size: 1.2rem;
-}
-
-.logoutButton:hover {
-  opacity: 1;
-  transition: opacity .3s;
-}
