--- src/firebase/LoginMenu.js
@@ -1,40 +1,0 @@
-import { useDispatch, useSelector } from "react-redux";
-import { useEffect, useState } from "react";
-import './loginMenu.css';
-import { useDatabase } from "../hooks/database";
-
-export const LoginMenu = () => {
-  const dispatch = useDispatch();
-  const {readUserCountries, readUserPermissionVisited} = useDatabase();
-  const currUser = useSelector(state => state.currUser.currUser);
-  const srcAvatar = (currUser) ? currUser.photoURL : null;
-  const userName = (currUser) ? currUser.displayName : null;
-  const [showMod, setShowMod] = useState(false);
-
-  useEffect(
-    () => {
-      if (currUser) {
-        readUserCountries(dispatch);
-        readUserPermissionVisited(dispatch);
-      }
-    }, [currUser]);
-
-  const changeLogoutOpen = () => {
-    setShowMod(true);
-  }
-
-  return (
-    <>
-      {(currUser) ?
-        <div className='menuAvatar'>
-          <img
-            className='imageAvatar'
-            src={currUser.photoURL}
-            alt={userName}
-          />
-        </div> :
-        <div className='loginMenu'>Login</div>
-      }
-    </>
-  );
-}
