+++ src/firebase/login.css
@@ -1,0 +1,45 @@
+.menuAvatar {
+  display: inline-block;
+  position: relative;
+  cursor: pointer;
+}
+
+.imageAvatar {
+  width: 2.5rem;
+  border-radius: 50%;
+  vertical-align: middle;
+  margin-left: 0.2rem;
+}
+
+.loginMenu {
+  font-size: 1.3rem;
+  vertical-align: middle;
+  color: #b1091d;
+  text-decoration: none;
+  cursor: pointer;
+  transition: color .5s;
+}
+
+.loginMenu:hover {
+  color: #f30e29;
+}
+
+.logoutMenu {
+  position: absolute;
+  transform: translate(-85%, 120%);
+}
+
+@media (max-width: 600px) {
+  .loginMenu {
+    margin: 0 0.5rem;
+    font-size: 1.25rem;
+  }
+}
+
+@media (max-width: 400px) {
+  .loginMenu {
+    margin: 0.2rem;
+    font-size: 1rem;
+    line-height: 1.75rem;
+  }
+}
