+++ src/hooks/database.js
@@ -2,13 +2,12 @@ import { useSelector } from 'react-redux';
 import { child, get, ref, set } from "firebase/database";
 import { database } from "../firebase/firebase";
 import { updateFavData } from "../redux/favCountriesSlice";
-import { setAllowShowVisited } from "../redux/loginUsersSlice";
 
 export const useDatabase = () => {
   const currUser = useSelector(state => state.currUser.currUser);
   const userId = currUser ? currUser.uid : null;
 
-  function writeUserCountries(countries) {
+  function writeUserData(countries) {
     if (userId) {
       set(ref(database, 'users/' + userId), countries);
     } else {
@@ -16,7 +15,7 @@ export const useDatabase = () => {
     }
   }
 
-  function readUserCountries(dispatch) {
+  function readUserData(dispatch) {
     if (userId) {
       const dbRef = ref(database);
       get(child(dbRef, `users/${userId}`)).then((snapshot) => {
@@ -36,33 +35,5 @@ export const useDatabase = () => {
     }
   }
 
-  function writeUserPermissionVisited(isAllow) {
-    if (userId) {
-      console.log("google: ", isAllow);
-      set(ref(database, `settings/${userId}/allowShowVisited/`), isAllow);
-    } else {
-      console.log('No auth !');
-    }
-  }
-
-  function readUserPermissionVisited(dispatch) {
-    if (userId) {
-      const dbRef = ref(database);
-      get(child(dbRef, `settings/${userId}/allowShowVisited`)).then((snapshot) => {
-        if (snapshot.exists()) {
-          const dataString = snapshot.val();
-          let isAllow = JSON.parse(dataString);
-          console.log("load: ", isAllow)
-          dispatch(setAllowShowVisited(isAllow))
-        } else {
-          console.log("No data available");
-        }
-      }).catch((error) => {
-        console.error(error);
-      });
-    }
-  }
-
-  return {writeUserCountries, readUserCountries, writeUserPermissionVisited, readUserPermissionVisited};
-  // return {writeUserData: writeUserCountries, readUserData: readUserCountries};
+  return {writeUserData, readUserData};
 }
