+++ src/redux/loginUsersSlice.js
@@ -3,7 +3,6 @@ import { createSlice } from '@reduxjs/toolkit';
 const initialState = {
   currUser: 0,
   idCurrUser: null,
-  isAllowShowVisited: true
 }
 
 export const loginUserSlice = createSlice({
@@ -16,16 +15,9 @@ export const loginUserSlice = createSlice({
         state.idCurrUser = action.payload.currUser.uid;
       }
     },
-    setAllowShowVisited: (state, action) => {
-      console.log("redux: ", action.payload);
-      state.isAllowShowVisited = action.payload;
-    }
   }
 });
 
-export const {
-  setCurrUser,
-  setAllowShowVisited
-} = loginUserSlice.actions;
+export const {setCurrUser} = loginUserSlice.actions;
 
 export default loginUserSlice.reducer;
